Add files via upload
In the Course.java file the date format used is that of jodatime which has to be added as a project dependency. This is just a simple program that add students and prompts them to enter the course they are doing the modules of that course and later prints out the list of students "Surname" then with their their course and modules. Also among to be printed is the modules and the respective list of courses associated with each course.